Fix scrolling speed when scrolling acceleration is set to zero
